requirements.txt
delorean==1.0.0
requests==2.18.4
parse==1.8.2
PyYAML==3.12
beautifulsoup4==4.6.0
feedparser==5.2.1
selenium==3.12.0
Pillow==5.1.0
xmltodict==0.11.0
PyPDF2==1.26.0
python-docx==0.8.6
